Main and Recursion Class Fixes
Added error to the userInput scanner so it now correctly takes inputs.
For the Recursion class, switched the order of the if statement to
correct it, and added a Recursion constructor in order to use the method
within the print statement in the main method. Changed the name of the
method to reflect this change.